@@ -1,18 +1,5 @@
-import "bootstrap/dist/css/bootstrap.css";
 import React from "react";
 import ReactDOM from "react-dom";
-import { BrowserRouter } from "react-router-dom";
-import App from "./App.js";
-import registerServiceWorker from "./registerServiceWorker";
-
-const baseUrl = document.getElementsByTagName("base")[0].getAttribute("href");
-const rootElement = document.getElementById("root");
-
-ReactDOM.render(
-  <BrowserRouter basename={baseUrl}>
-    <App />
-  </BrowserRouter>,
-  rootElement
-);
-
-registerServiceWorker();
+import App from "./components/App.js";
+import "./scss/app.scss";
+ReactDOM.render(<App />, document.getElementById(`app`));
